rename: 엔티티 Product를 Asset으로 변경
